requirements.txt
python-social-auth==0.2.13
PyJWT==1.4
html5lib==0.9999999
bleach==1.4.2
Django==1.9.1
PyYAML==3.10
django-sslserver==0.18
python-dotenv==0.3.0
dj-database-url==0.3.0
requests==2.9.1
django-jsonify==0.3.0

#bbp firewalled packages
-e ./packages/task-types-0.0.10
-e ./packages/bbp-services-0.0.11
-e ./packages/hbp-app-python-auth-0.0.4
-e ./packages/bbp-client-0.3.14
